requirements.txt
asgiref==3.2.7
autopep8==1.5.4
awsebcli==3.18.1
botocore==1.15.49
cement==2.8.2
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==3.0.7
django-ckeditor==5.9.0
django-cors-headers==3.5.0
django-crispy-forms==1.9.2
django-debug-toolbar==2.2
django-filter==2.3.0
django-heroku==0.3.1
django-import-export==2.3.0
django-js-asset==1.2.2
django-money==1.1
django-userforeignkey==0.3.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
docutils==0.15.2
et-xmlfile==1.0.1
future==0.16.0
gunicorn==20.0.4
idna==2.7
img2pdf==0.4.0
jdcal==1.4.1
jmespath==0.10.0
lxml==4.5.2
Markdown==3.2.2
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.3
pathspec==0.5.9
pdf2image==1.13.1
pdf2jpg==1.0
pikepdf==1.18.0
Pillow==7.1.2
psycopg2==2.8.6
py-moneyed==0.8.0
pycodestyle==2.6.0
PyJWT==1.7.1
pypiwin32==223
python-dateutil==2.8.0
python-decouple==3.3
python-docx==0.8.10
pytz==2020.1
pywin32==227
PyYAML==5.3.1
requests==2.20.1
semantic-version==2.5.0
six==1.11.0
sqlparse==0.3.1
tablib==2.0.0
termcolor==1.1.0
toml==0.10.1
urllib3==1.24.3
wcwidth==0.1.9
whitenoise==5.2.0
xlrd==1.2.0
xlwt==1.3.0
